Refactored helpers to reusable helpers file
